Refactor: move default unit converter things aroudn to facilitate use of optional import pint
