+++ pdtable/units/converter.py
@@ -1,0 +1,11 @@
+"""Default unit converter"""
+
+try:
+    from .pint import PintUnitConverter as DefaultUnitConverter
+except ImportError as err:
+    raise ImportError(
+        "Unable to import 'pint'. Please install 'pint' to use the default unit converter."
+    ) from err
+
+# Singleton, for convenience.
+convert_units = DefaultUnitConverter()
